// src/models/authority.enum.ts
///
/// Copyright © 2016-2021 The Thingsboard Authors
///
/// Licensed under the Apache License, Version 2.0 (the "License");
/// you may not use this file except in compliance with the License.
/// You may obtain a copy of the License at
///
///     http://www.apache.org/licenses/LICENSE-2.0
///
/// Unless required by applicable law or agreed to in writing, software
/// distributed under the License is distributed on an "AS IS" BASIS,
/// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
/// See the License for the specific language governing permissions and
/// limitations under the License.
///
export enum Authority {
    SYS_ADMIN = 'SYS_ADMIN',
    TENANT_ADMIN = 'TENANT_ADMIN',
    CUSTOMER_ADMIN = 'CUSTOMER_ADMIN',
    CUSTOMER_USER = 'CUSTOMER_USER',
    TENANT_USER = 'TENANT_USER',
    REFRESH_TOKEN = 'REFRESH_TOKEN',
    ANONYMOUS = 'ANONYMOUS'
}